+++ pkgs/async/lib/src/result/value.dart
@@ -2,13 +2,10 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'dart:async';
-
-import 'error.dart';
-import 'result.dart';
+part of 'result.dart';
 
 /// A result representing a returned value.
-class ValueResult<T> implements Result<T> {
+final class ValueResult<T> implements Result<T> {
   /// The result of a successful computation.
   final T value;